pyproject.toml
[project]
name = "kristen-race-director"
version = "0.1.0"
description = "Automation daemon for MultiViewer for F1 that dynamically rotates onboard camera windows during live races"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.24",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "structlog>=24.0",
    "mvf1>=0.1",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff>=0.1"]

[project.scripts]
race-director = "race_director.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["race_director"]
